Annotate testing of -dM -dD flags as requiring GCC frontend
This test requires the flags -dM -dD to be passed in a certain way.

The flag -dM prints macro definitions *instead of* normal output.
The flag -dD prints macro definitions *in addition to* normal
output.

GCC interprets this as -dD will re-enable normal output if it was
previously disabled (such as by -dM).
Clang interprets this as -dD will turn on macro definitions and not
touch normal output, leaving it disabled.